refactor(quiz): scope i18n to standalone widget
StandaloneQuizWidget is now an async server component (QuizWidget/index.tsx) that provisions its own client i18n context, subset to the quiz's own question strings, the widget UI keys, and the quiz-title key. The interactive widget moves to QuizWidgetClient.tsx behind the client boundary; client consumers (quizzes hub, stories, useQuizWidget types) import it directly and only server modules import the barrel.

Pages render the widget directly with no getMessages/pick/I18nProvider boilerplate, and client-shipped messages per quiz drop from ~106 KB (all of learn-quizzes + common) to ~5 KB. QUIZZES_PREFIXES and the use-cases layout learn-quizzes entry are removed; /quizzes/ keeps the full namespace as its primary namespace.
