+++ app/[locale]/privacy/page.tsx
@@ -1,4 +1,3 @@
-import { pick } from "lodash"
 import {
   ArrowLeftRight,
   ChartNoAxesCombined,
@@ -6,18 +5,13 @@ import {
   Lightbulb,
   Sparkle,
 } from "lucide-react"
-import {
-  getMessages,
-  getTranslations,
-  setRequestLocale,
-} from "next-intl/server"
+import { getTranslations, setRequestLocale } from "next-intl/server"
 
 import type { Lang, ToCItem } from "@/lib/types"
 
 import BigNumber from "@/components/BigNumber"
 import PathwayCard from "@/components/cards/pathway-card"
 import PageHero from "@/components/Hero/PageHero"
-import I18nProvider from "@/components/I18nProvider"
 import { Image } from "@/components/Image"
 import { Strong } from "@/components/IntlStringElements"
 import { StandaloneQuizWidget } from "@/components/Quiz/QuizWidget"
@@ -42,7 +36,6 @@ import VideoWatch from "@/components/Videos/VideoWatch"
 import { getAppPageContributorInfo } from "@/lib/utils/contributors"
 import { getMetadata } from "@/lib/utils/metadata"
 import { numberFormat, numberToPercent } from "@/lib/utils/numbers"
-import { getRequiredNamespacesForPage } from "@/lib/utils/translations"
 
 import PageJsonLD from "./page-jsonld"
 
@@ -57,10 +50,6 @@ const Page = async (props: { params: Promise<{ locale: Lang }> }) => {
   setRequestLocale(locale)
 
   const t = await getTranslations("page-privacy")
-
-  const allMessages = await getMessages({ locale })
-  const requiredNamespaces = getRequiredNamespacesForPage("/privacy/")
-  const messages = pick(allMessages, requiredNamespaces)
 
   const { contributors, lastEditLocaleTimestamp } =
     await getAppPageContributorInfo("privacy", locale)
@@ -454,9 +443,7 @@ const Page = async (props: { params: Promise<{ locale: Lang }> }) => {
         </Section>
 
         <Section id="quiz-section">
-          <I18nProvider locale={locale} messages={messages}>
-            <StandaloneQuizWidget quizKey="privacy" />
-          </I18nProvider>
+          <StandaloneQuizWidget quizKey="privacy" />
         </Section>
       </ContentLayout>
     </>
