+++ app/[locale]/quizzes/_components/quizzes.tsx
@@ -8,7 +8,7 @@ import ContentFeedback from "@/components/ContentFeedback"
 import { HubHero } from "@/components/Hero"
 import Github from "@/components/icons/github.svg"
 import MainArticle from "@/components/MainArticle"
-import QuizWidget from "@/components/Quiz/QuizWidget"
+import QuizWidget from "@/components/Quiz/QuizWidget/QuizWidgetClient"
 import QuizzesList from "@/components/Quiz/QuizzesList"
 import QuizzesModal from "@/components/Quiz/QuizzesModal"
 import QuizzesStats from "@/components/Quiz/QuizzesStats"
